+++ lib/src/base/types/Dynamic.vue
@@ -1,6 +1,6 @@
 <template>
-  <div>
-    <h4 class="dynamic-t">{{ props.linked_to ? formatName(props.linked_to, '/') : $t('Adding section') }}</h4>
+  <div class="sub-types dynamic-s">
+    <h4 :class="{'dynamic-t': isSideBarOpen === false}">{{ props.linked_to ? formatName(props.linked_to, '/') : $t('Adding section') }}</h4>
     <div v-if="globalSectionMode === true" class="mt-4">
       <div class="autoInsertRow">
         <div>
@@ -60,7 +60,11 @@ export default {
     basePath: {
       type: String,
       default: ''
-    }
+    },
+	isSideBarOpen: {
+	  type: Boolean,
+	  default: false
+	}
   },
   components: {
     GlobalReferences
@@ -235,5 +239,8 @@ export default {
 }
 .instanceInput {
   width: 350px;
+}
+.dynamic-s {
+  text-align: center;
 }
 </style>